WIP: refactoring code. Moving many util methods into classes.
